# internal/manifest: incrementally generate L0Sublevels for L0 additions