NewL0Sublevels is an expensive method to call, and it's called every time
there's a change to L0. This commit adds a faster, incremental generation
method for cases when there are only additions to L0 and no deletions from
it. This lets us reuse most of the slices in NewL0Sublevels and merge new
fileIntervals into it instead of starting from scratch.

Improvement in load times for `MANIFEST_import`:

```
name                             old time/op  new time/op  delta
ManifestApplyWithL0Sublevels-12   27.1s ± 4%    4.0s ± 9%  -85.05%  (p=0.000 n=10+10)
```

Fixes #1406.