@itsbilal 2022-01-21
Loading the `MANIFEST_import` manifest in testdata (taken from a large TPCC import)is now 85% faster with this optimization:

```
name                             old time/op  new time/op  delta
ManifestApplyWithL0Sublevels-12   27.1s ± 4%    4.0s ± 9%  -85.05%  (p=0.000 n=10+10)
```

And that reduction is with all the overhead involved in reading the manifest and applying version edits. The actual improvement in `L0Sublevels` generation time is definitely well above 90% if it is able to improve a coarse benchmark that much.