@sumeerbhola 2022-02-03
Future PR is fine by me too, for the randomized test. We should definitely prioritize it so that it happens before CockroachDB v22.1.