@itsbilal 2022-02-04
___
*[internal/manifest/l0_sublevels.go, line 463 at r3](https://reviewable.io/reviews/cockroachdb/pebble/1465#-Mun7Toj4_cthOXZ1ZnQ:-Mv5KvqwCwd5KPqlx1C_:b-6z45ky) ([raw file](https://github.com/cockroachdb/pebble/blob/88795ab5ed07ee20ba4b5764f174025fa4784bb3/internal/manifest/l0_sublevels.go#L463)):*
<details><summary><i>Previously, itsbilal (Bilal Akhtar) wrote…</i></summary><blockquote>

Done.  Will add the randomized test in a future PR, but I've added an invariant check in `TestL0Sublevels`. 
</blockquote></details>

I'm happy to merge as-is, but will give Sumeer the chance to scrutinize the new `estimateBytes` logic as it's a little different.


<!-- Sent from Reviewable.io -->